requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
idna==2.7
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.6.4
six==1.11.0
